renamed ripple to Ripple
