+++ src/components/button/button.js
@@ -1,4 +1,4 @@
-import Ripple from './ripple';
+import Ripple from './Ripple';
 
 const buttons = document.querySelectorAll('.js-ripple');
 buttons.forEach(button => new Ripple(button));
